Make trace of fetch-like functions an option.
Added a trace_fetch parameter to the main class getter func. The idea
is to trace only execute and executemany by default, and keeping all
the fetch... functions as-is (untraced) by default, as most users will
probably only want the execute-like funcs results, as fetch should be
"fast enough" in most cases.