[distributed sampling] Introduced priority sampler.
A new priority sampler is available, unlike the other sampler which
is only based on a local rate, this one uses feedback from the agent
to decide wether a given trace should be sampled or not.

It is still possible to use the historical sampler also, this one
operates later in the pipeline. Whenever a trace should be sampled,
an attribute is set which is a message for the downstream backend that
"this should be kept, no matter what" and conversely, the same
attribute can be set to state that "this should be used for stats, but
not be kept as a trace". The decision is made very early in the
pipeline, when the root span is created. The attribute is then
propagated through all calls, local or remote.

By default, this feature is disabled, is has to be explicitely
activated by passing 'priority_sampling = True' to configure().