[aiohttp] Fix race condition in testing
The current code only yield once the thread have started, though there's a
little chance that one of the thread did not had time to actually sent its
request yet. That means that the requests will never be served and the thread
will fail.

This patch fixes this by making sure to do pass the execution to the asyncio
loop between each join, just in case a request needs to be handled.

Fixes #876