Add a PeriodicWorker base class for periodic tasks (#934)
There is at least 2 places (writer and runtime metric collectors) where the
code needs to start worker threads doing regular tasks.

This patch merge the existing code in a simple PeriodicWorker base class that
can be enhanced with custom code. It has the upside of using a
`threading.Event` rather than a sleep timer, making the exit code running as
soon as `stop()` is called.