tracer: fix configure(collect_metrics) argument
- The current code does not allow to disable collect_metrics by passing `False`
  to it.
- The recent change from d5e843dfd3c024b357a159e890b44921965fcdf9 made the hack
  in test_runtime_metrics to change the FLUSH_INTERVAL a noop, making the test
  taking 20 seconds to complete.

This patch fixes both issues by allowing the `collect_metrics` arguments to be
a boolean or a float representing the interval in seconds to use to collect the
runtime metrics. This makes sure users can disabled runtime metrics at any
time, and fixes the test for running in a less than a second.