[contrib/cassandra] Handle batched bound statements in python3 (#1062)
* [contrib/cassandra] Handle batched bound statements in python3

When a bound statement is batched, the query string it uses is it's
query_id which is bytestring (and not an encoded one, just a raw bytes
sequence) so when it is attempted to be joined using a unicode string
(in python3) it raises an exception due to mixed types during query
sanitization.  If the prepared flag is set to `True` on the statements
list, don't include the query_id.

* Improve context in code

Add in a comment explaining the purpose of the filter on query
concatenation for batch statements and document the fields of the
_statements_and_parameters tuples (along with a link to the source).

Added an assertion around tag values to the test as well.