feat(profiling): retry on upload error (#1376)
This adds a retry mechanism to the HTTP exporter for the profiler. It'll retry
exponentially with some randomized jitter to avoid collisions from multiple
clients if the service goes down.

This adds a dependency on tenacity >=5 which was release 2 years ago.

This changes also retry on IOError and OSError, which includes any kind of
socket error (including timeout) to make things more robust.