fix(profiling): identify Python main thread properly with gevent (#1445)
When gevent is running and monkey-patches the threading module, the stack
collector struggles to identify the MainThread. This is due to gevent replacing
the native thread with greenlet based ones.

This patch leverages the original threading.get_ident() function to store the
main thread id and stores it for future reference, even if the application has
been monkey-patched.