fix(profiling): multi-threading/gevent issues with ThreadLinkSpan (#1485)
* fix(profiling/stack): fix ThreadLinkSpan locking with gevent

The current code uses a threading.Lock object, which in gevent can be patched
with a gevent-lock. Trying to synchronize our real OS thread with the Python
main thread using a gevent-lock does not work and crashes the program
altogether.

This makes sure we use an original threading.Lock to synchronize the 2 OS threads.

* fix(profiling/stack): replace WeakSet by set

This replaces the usage of WeakSet by a set using directly weakref.
This makes sure we control the entire lifespan of our set and refs, otherwise
they might disappear under us if e.g. the GC runs while iterate over it.