feat(profiling): restart the Profiler on fork() (#1535)
* refactor(profiling): add Profiler.copy

* refactor(profiling): remove recorders attribute

* feat(profiling): restart the Profiler when fork() happens

This adds an indirection layer for the Profiler object in order to be able to
restart the Profiler on fork.

The Profiler class is now a very high-level profiling object that represents a
whole life-cycle of profiling. The ProfilerInstance is now the actual instance
doing the profiling; this one can be recreated whenever we need: the first use
case being on fork().