fix(profiling): fix thread join() hang after fork() with gevent (#1877)
In gevent when profile_childre is True, the child process will try to stop the
parent profiler. This can hang forever since the thread is not running: it will
never set the `quit` value to True.

To avoid this, we use `self.is_alive` to check if the thread is alive; on fork,
the thread is marked as dead by Python via the _reset_internal_locks() method.

This is a fix similar to ab4e17b87caeb6dd8f6dd74c99d3cde5e53c4e8d

In order to `is_alive` to work, the thread now set the Thread._started event
correctly. It's set in __bootstrap normally, but this uses Locks in Python 3
which are now gevent-Lock, so we can't use this method. We use our own variable
and a busy loop to make sure the thread booted.

This can avoid a race condition when the process fork just after the thread
starts and where it'd not be in threading._active. You never know!

Fixes #1414