test(writer): test writer against real API endpoint (#2023)
This ports the existing test from test_api to the new writer code to make sure
we handle those cases correctly.

This only test the lower layer for now; more test to see if errors are
correctly caught and handled cannot be done easily as the datadog logger has a
rate limit that can't be easily disabled.