Fix requests service naming (#2273)
The service naming for the requests integration used the wrong type of
service configuration. requests spans should not inherit from parent
spans as requests are made to external services.

In addition to the above fix support is added to the integration for the
global configuration using the config API or environment variables.

The docs are also rewritten to be consistent with our newer
integrations.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>