perf: cache http data (#2290)
* perf: cache http data

This change aims at improving performance of integrations that juggle
HTTP data, like headers and error codes. For this to work, the
assumption that customisation is set once at start-up, before any of the
cached functions are called, **must** hold true. Otherwise, this is
bound to become a subtle source of bugs.

* Use dedicated cache for error code range parsing

Some code has been reshuffled around to avoid circular imports

* ensure per-instance method caching

* add type hints and more caching

* add tests for the new caching utils

* reuse caching code for methods

* pytest plugin to invalidate caches between tests

* add docstring to cached decorator

* fix docstring typo

Co-authored-by: Kyle Verhoog <kyle@verhoog.ca>

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>
Co-authored-by: Brett Langdon <brett.langdon@datadoghq.com>
Co-authored-by: Kyle Verhoog <kyle@verhoog.ca>