Adds DD_TRACE_AGENT_TIMEOUT_SECONDS (#2508)
For sending apm metrics to localhost, the defaults of 2.0 is enough. However, for setup where datadog agents are exposed over https, the timeout of 2.0 is not sufficient. Such kind of setup is used for non-prod environment for devs to test their apm/stats configuration without having to install ddagent on every non-prod machines or generate multiple ddagent key for each dev to test themselves.

Co-authored-by: Brett Langdon <me@brett.is>
Co-authored-by: brettlangdon <brett.langdon@datadoghq.com>