Add forksafe threading objects (#2511)
* refactor(forksafe): add forksafe threading Lock classes

This adds a wrapper around the Lock, RLock and Event class in order to provide
a fork-safe equivalent of the existing classes.

This objects can be re-used after a fork safely as they are reset.

* refactor(profiling): leverage forksafe.Lock in DoubleLock

This leverages the forksafe Lock to prevent dead-lock after fork.
This makes sure that when the parent-Profiler is stopped in the child process,
it can actually do it without trouble since the recorder DoubleLock is sure to
be freed.

* fix(service): use forksafe Lock and Event objects

This replaces the existing and regular Lock and Event classes from the Python
threading by their forksafe alternative, making sure that a Service can be
safely stopped or a PeriodicThread joined once a process has forked.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>