fix(tracer): deprecate reuse of shut down tracer (#2541)
* refactor(tracer): make shut down tracer unusable

Currently, a tracer that has been shut down can be reused by attempting
to trace again. However, at this point the atexit hook has been
unregistered and doesn't get registered again. Adding a check to
Tracer.trace would cause overheads on a hot path, so this change makes
a tracer that has been shut down unusable for further tracing, unless it
is recreated.

* fix(tracer): deprecate reuse of shut down tracer

Reusing a tracer that has been shut down without an explicit restart is
problematic. This change adds a deprecation warning that is emitted
whenever a tracer that has been shut down is reused for tracing and also
ensures that the atexit hook is re-registered when this happens.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>
Co-authored-by: Julien Danjou <julien.danjou@datadoghq.com>