test(profiling): rework gevent task ignore test (#2558)
The existing test works fine in 99% of the case, but since the fib() call is a
bit too short and the while loop timing a bit too fast, in some circumstances,
it might take several minutes for the profiler to catch it.

This extends the time needed to run fib() to have a better change to catch the
task, and refactor the opposite test so the code is shared across the both
positive and negative tests.

I ran this test in a loop a few hundreds of time and it never failed. 🤞

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>