fix(django): ensure expected string type for tags (#2465)
* fix(django): ensure expected string type for tags

PR #2174 introduced a regression in Django whereby some tags are being
assumed to be of string type and set directly using `_set_str_tag` for
better performance. Issue #2418 has highlighted that this assumption is
violated for the `template_name` tag, which could take values of type
`PosixPath`.

This change guards against this case by trying to coerce the value to a
string before handing it over to `_set_str_tag`.

Fixes #2418.

* add PosixPath template name test

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>