Context copy improvements (#2585)
* Context copy improvements

Each span needs its own context copy as the `span.context` property
returns the trace context specialized with the span's id.

The lock is removed from `context._with_span` as the copied references
are not volatile.

* Handle possibly none context

The span._context can possibly be None but only when not using the main
tracer APIs. Add a safeguard in case the main APIs are not used.

* Don't set none attributes on initialize

This prevents needless lock acquisition when initializing context
objects.