refactor(writer): move to a buffered encoder (#2422)
* refactor(writer): buffer traces instead of bytes

The current implementation of the AgentWriter relies on a buffer
structure that stores encoded versions of traces in order to comply with
payload size limits. As a drawback, the encoder needs a "join_encoded"
method to stich together traces that have already been encoded.

In light of upcoming work to support the v0.5 format for traces payload,
this change proposes to make the buffer store unencoded traces and have
their sizes estimated at best. The encoding is then done once by the
AgentWriter worker thread, just before uploading the result to the
Datadog agent.

The size estimation becomes a very delicate aspect. Trying to be too
accurate introduces noticeable overheads (estimated around 1.22% for a
fairly accurate trace size heuristics). The  implementation in this
change proposes a more approximate trace size guessing which, in terms
of performance, appears to be on par with the current implementation.

* cythonise the heavy trace size estimator

With this change, not only trace size estimates are quite accurate (the
writer tests pass in their original form), but a slight performance gain
can also be noticed (~0.6)%

* fix/add typing

* add ticker to allow other threads to run

* compute exact trace size

* shrink encoder API

* introduce buffered encoders

* tidy up packer

* reintroduce joining logic

With the new refactor, the joining logic saves computing the trace size;
however, when a trace doesn't fit into the buffer, the encoding would be
wasted (like in the previous implementation).

* add buffer limit tests

Co-authored-by: Brett Langdon <brett.langdon@datadoghq.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>
Co-authored-by: Tahir H. Butt <tahir.butt@datadoghq.com>
Co-authored-by: Kyle Verhoog <kyle@verhoog.ca>