fix(ddtrace-run): do not handle errors wrongly (#2146)
* fix(ddtrace-run): do not handle errors wrongly

ddtrace-run tries to be smart and handle all `OSError` and `PermissionError`
with the same code.

This is actually worse than doing nothing. I had the script printing the error
"executable does not have executable permissions" whereas the real problem was
an `OSError` with errno set to ENOEXEC, not EACCESS — meaning the shebang was
missing in the script, while it did have the +x bit.

Trying to print the same error message for all possible errno values is just
confusing for users. At this stage, it's easier to leave `execl` raises the
error back to the user so they can understand what might be wrong.

* Linting and formatting

* Python 2 test compatibility

Co-authored-by: Kyle Verhoog <kyle@verhoog.ca>
Co-authored-by: Kyle Verhoog <kyle.verhoog@datadoghq.com>