test(profiling): enhance gevent thread hub/task detection (#2990)
This test regularly fails because the first events from the stack collector are
from the gevent thread booting up, while we expect to get frames from the
thread running.

This changes the test to ignore events where we see the thread boot time, and
only look for when the thread actually runs. If we find it once, we know we're
good.