feat(profiling/gevent): supports tracking tasks in gevent < 1.3 (#2929)
This implements a greenlet tracer similar to the one that gevent 1.3 ships. It
is actually simpler than the one gevent provides as we don't need as many
features, so it should be _slightly_ faster.

This makes sure we can provide the same amazing experience whatever the gevent
version being used might be.

Co-authored-by: Kyle Verhoog <kyle@verhoog.ca>
Co-authored-by: Brett Langdon <brett.langdon@datadoghq.com>