fix(internal) cdef Python exceptions being ignored (#3008)
* fix(internal) cdef Python exceptions being ignored

In a cdef function `raise` statements will create a Python
exception, add it to the exception stack, and then immediately
return. However, the caller will not explicitly check to see if
any exception has occurred unless an exception return value has
been provided.

Until now these cdef functions when they encountered a `raise`
would push the Python exception, and then return the default
`int` value of `0`. Which based on how we use the return values
`0` means success.

This means that anytime we encounter an exception in our encoder
we potentially ignore any exceptions and still append the
partially encoded trace into the buffer.

With this change we are standardizing around `except? -1` which
tells the caller that their _might_ be a Python exception if the
return value is `-1`. We do this because the `msgpack_*`
functions we call will also return `-1` on error, but will not
create a Python exception.

* add release note

* Update integration tests

* remove unused import

* set file encoding

* Update releasenotes/notes/fix-encoding-exceptions-c12900b38741d2bf.yaml

Co-authored-by: Tahir H. Butt <tahir.butt@datadoghq.com>

Co-authored-by: Tahir H. Butt <tahir.butt@datadoghq.com>