fix reno pre_release_tag_re pattern
We have been having issues with how reno collapses pre-release notes into official release versions.

This change fixes a bug in how reno decides to collapse pre-release notes.

If we have the tags:

- v0.56.0
- v0.56.0rc7
- v0.56.0rc6
- ...

When reno iterates over these versions it will properly identify `v0.56.0rc7` as a pre-release, but then will identify `.0rc7` as the pre-release identify and strip it from the tag. So we end up with `v0.56` and it then will not collapse it because it cannot find a release for `v0.56`.

This change moves the `\v.\d+` pattern to outside the `pre_release` matching group in the regular expression so we can properly strip just `rc#` from the tag.