fix: support django channels >= 3.0 (#2956)
Currently ddtrace is not working with daphne #2945. This is due to how django channels configures consumers.

Instead of passing a view to django.urls.path and django.urls.re_path django channel's URLRouter expects an asgi application. This causes an application to fail on startup after we incorrectly wrap an asgi application as a view.

To fix this issue this change checks if the asgi application callback in URLPatterns is wrapped as a views. If so, we unwrap the asgi application. This approach should fix the issue described above.

Co-authored-by: Brett Langdon <me@brett.is>
Co-authored-by: Kyle Verhoog <kyle@verhoog.ca>