fix(ci): Re-enable spell checking for PR release notes (#3156)
In #3085 when we moved to manually specifying release branches
we stopped including any release notes on the current branch
in the docs build.

This means that spell checking for a release note won't occur
until it is merged into a release branch which is far too late.

With this change we will include the full change log entry list
for the current branch when we are using the 'spelling'
builder.