fix: decode bytes strings when using JSON encoder (#3117)
## Commit Message

{{title}}

If a `bytes` string is used in span metadata and you are using
the lambda layer/`LogWriter` which uses the `JSONEncoder`
you will get an exception.

This change turns our base `JSONEncoder` class into a subclass
of `json.JSONEncoder` so we can overwrite `JSONEncoder.default`
and properly decode `bytes`.

Subclassing from `json.JSONEncoder` means we don't need to define
an `encode` method on the base class.

Fixes #3115