fix(telemetry): ensure integrations are queued on startup (#3471)
Currently integrations are not sent to the trace agent (telemetry is enabled after patching occurs). 

This change ensures integrations are queued when ``patch`` or ``patch_all()`` is called. 

If telemetry is never enabled we will queue configuration data for up to 64 integrations and these values will persist in memory. The cost of each integration entry is ~100 bytes.

## Checklist
- [ ] Added to the correct milestone.
- [ ] Tests provided or description of manual testing performed is included in the code or PR.
- [ ] Library documentation is updated.
- [ ] [Corp site](https://github.com/DataDog/documentation/) documentation is updated (link to the PR).