chore(ci): use more parallelism (#3573)
Some of the jobs were using parallelism 6 but we run with 7 versions of
Python so some parallel runners run two versions of Python. By bumping
the parallelism to 7 we effectively half the job duration of profiling,
opentracing, kombu and celery.

We missed this regression when we added support for Python 3.10.

This cuts down our cI time from ~40 mins to ~25 mins.