fix(stack,asyncio): add missing CPU time in asyncio apps (#3576)
The current asyncio code does not report CPU time for the main thread when
there are tasks running.

The code expects one of the task to be the main thread; this is only true for
gevent. Asyncio does not have this behaviour, which means we only can account
wall time for tasks.

The CPU time must be assigned to the regular main thread, just like we'd do for
a non-async program.

Fixes #3572