fix(tracer/writer): lock the connection (#3578)
* fix(tracer/writer): lock the connection

There existed a critical section with the connection object on writer as
the `periodic` logic of the writer could be invoked from the periodic
writer thread as well as from other threads calling `periodic`
(`tracer.flush()` or `tracer.shutdown()`). The critical section would
result in corrupted http connections. This bug was introduced in #2999.

The solution is to lock the connection when used. I don't suspect that
this would lead to any performance degradation for the vast majority of
users since there is little overhead added to the writer. A degradation
would occur if contending threads were calling `periodic()` as now the
connection is a shared resource.

* Add comment for connection lock