chore(internal): add HTTP connector utility (#3619)
This change adds an HTTP connector utility that wraps a Python
connection object with a context manager for convenience and reuse.
The context manager essentially closes over the connection information
so that new connections to the same socket can be created by simply
calling the result of the connector helper. Examples are provided by the
new test cases.