feat(internal): add span sampling rule api (#3918)
## Description
The span sampling rule api is part of the single span sampling initiative. It takes in service and name patterns, sampling_rate, and a rate limit. From there it uses those to evaluate whether or not we should sample a span if it's not already being sampled by the normal trace sampler. If we should sample it, then we give it unique tags, which the trace-agent will recognize and keep the single span with those tags.

## Motivation
This is required in order for us to then apply the rules to the spans in a span processor.

## Design 
Described above.

## Testing strategy
Tested that the rule adds tags correctly. Although could use assistance if we want to tag that a certain percent of spans are tagged at a given sampling rate or something along those lines. Also tested that the rate limiter works when set to 0 (don't allow any spans through), but have not tested how exactly it reduces those spans. I assume that to some degree, since both the sampler and the rate limiter have their own tests, this may be unnecessary.  


## Checklist
- [ ] Title must conform to [conventional commit](https://github.com/conventional-changelog/commitlint/tree/master/%40commitlint/config-conventional).
- [ ] Add additional sections for `feat` and `fix` pull requests.
- [ ] Ensure tests are passing for affected code.
- [ ] [Library documentation](https://github.com/DataDog/dd-trace-py/tree/1.x/docs) and/or [Datadog's documentation site](https://github.com/DataDog/documentation/) is updated. Link to doc PR in description.


## Reviewer Checklist
- [x] Title is accurate.
- [x] Description motivates each change.
- [x] No unnecessary changes were introduced in this PR.
- [x] PR cannot be broken up into smaller PRs.
- [x] Avoid breaking [API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces) changes unless absolutely necessary.
- [x] Tests provided or description of manual testing performed is included in the code or PR.
- [x] Release note has been added for fixes and features, or else `changelog/no-changelog` label added.
- [x] All relevant GitHub issues are correctly linked.
- [x] Backports are identified and tagged with Mergifyio.
- [ ] Add to milestone.